Refactor transaction module and update import paths
